app.py
import streamlit as st
import time
import re
import json
from docx import Document
from io import BytesIO

# 导入你之前的核心逻辑函数（此处为示意，需确保函数在同一目录下或文件中）
# from ai_driven_processor import get_structured_blocks, get_detailed_analysis, load_answers

st.set_page_config(page_title="EduParser - 智能解析", layout="wide")

# 自定义 CSS 样式以接近 UI 图片
st.markdown("""
    <style>
    .main { background-color: #fff5f5; }
    .stButton>button { background-color: #e63946; color: white; width: 100%; border-radius: 5px; }
    .upload-card { border: 1px solid #f0f0f0; padding: 20px; border-radius: 10px; background: white; text-align: center; }
    </style>
    """, unsafe_allow_html=True)

# 顶部导航模拟
st.image("https://via.placeholder.com/150x50?text=EduParser", width=150)  # 替换为你的Logo

st.title("Welcome to EduParser!")
st.write("Generate Detailed Exam Solutions with Ease")

# 主体布局：左右两个上传框
col1, col2 = st.columns(2)

with col1:
    st.markdown('<div class="upload-card">', unsafe_allow_html=True)
    st.subheader("Upload Exam Paper")
    exam_file = st.file_uploader("Choose Word File", type=['docx'], key="exam")
    st.markdown('</div>', unsafe_allow_html=True)

with col2:
    st.markdown('<div class="upload-card">', unsafe_allow_html=True)
    st.subheader("Upload Answer Key (Optional)")
    ans_file = st.file_uploader("Choose TXT File", type=['txt'], key="ans")
    st.markdown('</div>', unsafe_allow_html=True)

# 中间生成按钮
if st.button("Generate Solutions", use_container_width=True):
    if exam_file:
        # 1. 初始化进度条
        progress_bar = st.progress(0)
        status_text = st.empty()

        # 2. 模拟/执行 AI 处理逻辑
        status_text.text("Analyzing and Generating Solutions, Please Wait...")

        # 步骤 A: 结构化识别 (示例进度 30%)
        # blocks = asyncio.run(get_structured_blocks(client, full_text))
        progress_bar.progress(30)
        time.sleep(1)  # 模拟耗时

        # 步骤 B: 循环解析 (示例进度每题增加)
        progress_bar.progress(65)
        time.sleep(1)

        # 3. 完成状态
        progress_bar.progress(100)
        st.success("Analysis Complete! Your Solution is Ready.")

        # 4. 下载区域
        st.markdown("---")
        st.subheader("Download Your Solution")
        # 假设 processed_docx 是你生成的 BytesIO 对象
        st.download_button(
            label="Download File",
            data=b"Your processed content",  # 这里替换为生成的文档二进制流
            file_name="Solution_Report.docx",
            mime="application/vnd.openxmlformats-officedocument.wordprocessingml.document"
        )
    else:
        st.error("Please upload an exam paper first!")

# 侧边栏：余额与充值
with st.sidebar:
    st.header("Top Up Your Account")
    st.markdown("### Current Balance: **¥120**")
    if st.button("Recharge Now"):
        st.info("Recharge feature coming soon...")