requirements.txt
streamlit
httpx
python-docx
tenacity